django-challenge/requirements.txt
Django==1.6.5
PyYAML==3.11
South==1.0
argparse==1.2.1
coverage==3.7.1
django-coverage==1.2.4
selenium==2.47.1
wsgiref==0.1.2
